% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_step_SIR_multihosts.R
\name{run_step_SIR_multihosts}
\alias{run_step_SIR_multihosts}
\title{Compute the change of the SIR multihosts system in one time-step}
\usage{
run_step_SIR_multihosts(t, state, parameters)
}
\arguments{
\item{t}{Current time}

\item{state}{Current state of the SEIR system}

\item{parameters}{Value of the epidemiological parameters:
\itemize{
\item tau Competences
\item b Birth rate
\item d Death rate
\item sigma Infectiosity period
\item alpha virulence
\item phi Number of contacts
}}
}
\value{
A list with a single element containing the derivative for the
SEIR system, as needed for \code{\link[deSolve:ode]{deSolve::ode()}}
}
\description{
Compute the change of the SIR multihosts system in one time-step
}
